Refactor surveys modules for complexity gate
