+++ packages/plugin-surveys/src/index.ts
@@ -1,15 +1,6 @@
 export { pluginSurveys } from './plugin.js';
-export {
-  pluginSurveysActionDefinitions,
-  pluginSurveysActionIds,
-  pluginSurveysExportActionId,
-  pluginSurveysModerationActionId,
-  pluginSurveysModuleIam,
-  pluginSurveysPermissionDefinitions,
-} from './plugin.js';
-export { pluginSurveysTranslations } from './plugin.translations.js';
 export { SURVEYS_CONTENT_TYPE } from './surveys.constants.js';
-export { deleteSurvey, getSurvey, listSurveys } from './surveys.api.js';
+export { deleteSurvey, listSurveys } from './surveys.api.js';
 export { SurveyCreatePage, SurveyEditPage } from './surveys.pages.js';
 export type {
   SurveyContentItem,
