+++ packages/plugin-surveys/src/plugin.translations.content.en.ts
@@ -1,0 +1,92 @@
+export const pluginSurveysContentEnTranslations = {
+  messages: {
+    createPendingHint:
+      'This section is already visible, but it will only be populated with data after the first save.',
+    sectionPlaceholder:
+      'The dedicated fields for this section follow in the next implementation steps.',
+    historyPlaceholder: 'The history will appear here once the survey has been created.',
+    unlimitedScheduleHint: 'Without an end date, the survey remains open-ended.',
+    targetAreasEmpty: 'There are currently no target areas available.',
+    metadataCreateHint: 'Metadata appears after the survey has been saved for the first time.',
+    freeTextQuestionHint: 'Free-text questions do not require answer options.',
+    optionSectionHint: 'Answer options remain flat sections inside the question.',
+    moderationEmpty: 'There are currently no free-text responses to moderate for this survey.',
+    moderationEmptyQuestion: 'There are currently no free-text responses for this question.',
+    deleteQuestionTitle: 'Delete question',
+    deleteQuestionDescription: 'Do you really want to delete this question?',
+    deleteOptionTitle: 'Delete option',
+    deleteOptionDescription: 'Do you really want to delete this answer option?',
+    deleteFreeTextTitle: 'Delete free-text response',
+    deleteFreeTextDescription: 'Do you really want to delete this free-text response?',
+    resultsQuestionEmpty: 'There are currently no results for this question.',
+    resultsFreeTextEmpty: 'There are no free-text responses for this question.',
+    resultsExportWithoutFreeText: 'Result export without free-text responses.',
+    resultsExportWithFreeText: 'Result export including free-text responses.',
+  },
+  validation: {
+    titleRequired: 'Please provide a title.',
+    questionTitleRequired: 'Please provide a question title.',
+    optionTitleRequired: 'Please provide an option title.',
+  },
+  fields: {
+    title: 'Title',
+    shortDescription: 'Short description',
+    description: 'Description',
+    status: 'Status',
+    startAt: 'Start',
+    endAt: 'End',
+    isAnonymous: 'Anonymous participation',
+    showResultsInApp: 'Show results in app',
+    resultVisibility: 'Result visibility',
+    privacyNotice: 'Privacy notice',
+    transparencyNotice: 'Transparency notice',
+    questionTitle: 'Question title',
+    questionDescription: 'Question description',
+    questionType: 'Question type',
+    questionRequired: 'Required question',
+    optionTitle: 'Option title',
+    optionEnablesFreeText: 'Allow free text',
+    freeTextExcerpt: 'Free text',
+    freeTextCreatedAt: 'Received',
+    freeTextStatus: 'Publicly visible',
+    freeTextOpenOverlay: 'Open full text',
+    freeTextOverlayText: 'Full text',
+    freeTextOverlayStatus: 'Status',
+    freeTextOverlayCreatedAt: 'Timestamp',
+    summaryParticipationCount: 'Participations',
+    summarySubmissionCount: 'Submissions',
+    summaryQuestionCount: 'Questions',
+    summaryStatus: 'Status',
+    optionVotes: 'Votes',
+    optionPercentage: 'Share',
+    freeTextSection: 'Free-text responses',
+    targetAreas: 'Target areas',
+    targetAreasSearch: 'Search target area',
+    targetAreasSearchPlaceholder: 'Select target area',
+    createdAt: 'Created',
+    updatedAt: 'Updated',
+    publishedAt: 'Published',
+    archivedAt: 'Archived',
+    statusOptions: {
+      draft: 'Draft',
+      active: 'Active',
+      archived: 'Archived',
+    },
+    resultVisibilityOptions: {
+      none: 'Never',
+      afterSubmission: 'After submission',
+      afterSurveyEnd: 'After survey end',
+    },
+    questionTypeOptions: {
+      singleChoice: 'Single choice',
+      multipleChoice: 'Multiple choice',
+      freeText: 'Free text',
+      singleChoiceWithText: 'Single choice with free text',
+      multipleChoiceWithText: 'Multiple choice with free text',
+    },
+    freeTextStatusOptions: {
+      internal: 'Internal',
+      public: 'Public',
+    },
+  },
+} as const;
