+++ packages/plugin-surveys/src/plugin.translations.content.ts
@@ -1,0 +1,92 @@
+export const pluginSurveysContentTranslations = {
+  messages: {
+    createPendingHint:
+      'Dieser Bereich ist bereits sichtbar, wird aber erst nach dem ersten Speichern mit Daten gefüllt.',
+    sectionPlaceholder:
+      'Die fachlichen Felder dieses Bereichs folgen in den nächsten Umsetzungsabschnitten.',
+    historyPlaceholder: 'Die Historie erscheint hier, sobald die Umfrage bereits angelegt wurde.',
+    unlimitedScheduleHint: 'Ohne Enddatum bleibt die Umfrage unbefristet.',
+    targetAreasEmpty: 'Es stehen derzeit keine Zielgebiete zur Auswahl.',
+    metadataCreateHint: 'Metadaten erscheinen nach dem ersten Speichern der Umfrage.',
+    freeTextQuestionHint: 'Freitextfragen benötigen keine Antwortoptionen.',
+    optionSectionHint: 'Antwortoptionen bleiben als flache Abschnitte innerhalb der Frage.',
+    moderationEmpty: 'Für diese Umfrage liegen derzeit keine Freitextantworten zur Moderation vor.',
+    moderationEmptyQuestion: 'Zu dieser Frage liegen derzeit keine Freitextantworten vor.',
+    deleteQuestionTitle: 'Frage löschen',
+    deleteQuestionDescription: 'Soll diese Frage wirklich gelöscht werden?',
+    deleteOptionTitle: 'Option löschen',
+    deleteOptionDescription: 'Soll diese Antwortoption wirklich gelöscht werden?',
+    deleteFreeTextTitle: 'Freitextantwort löschen',
+    deleteFreeTextDescription: 'Soll diese Freitextantwort wirklich gelöscht werden?',
+    resultsQuestionEmpty: 'Für diese Frage liegen derzeit keine Ergebnisse vor.',
+    resultsFreeTextEmpty: 'Für diese Frage liegen keine Freitextantworten vor.',
+    resultsExportWithoutFreeText: 'Ergebnisexport ohne Freitextantworten.',
+    resultsExportWithFreeText: 'Ergebnisexport inklusive Freitextantworten.',
+  },
+  validation: {
+    titleRequired: 'Bitte einen Titel angeben.',
+    questionTitleRequired: 'Bitte einen Fragetitel angeben.',
+    optionTitleRequired: 'Bitte einen Optionstitel angeben.',
+  },
+  fields: {
+    title: 'Titel',
+    shortDescription: 'Kurzbeschreibung',
+    description: 'Beschreibung',
+    status: 'Status',
+    startAt: 'Start',
+    endAt: 'Ende',
+    isAnonymous: 'Anonyme Teilnahme',
+    showResultsInApp: 'Ergebnisse in der App anzeigen',
+    resultVisibility: 'Ergebnisfreigabe',
+    privacyNotice: 'Datenschutzhinweis',
+    transparencyNotice: 'Transparenzhinweis',
+    questionTitle: 'Fragetitel',
+    questionDescription: 'Fragebeschreibung',
+    questionType: 'Fragetyp',
+    questionRequired: 'Pflichtfrage',
+    optionTitle: 'Optionstitel',
+    optionEnablesFreeText: 'Freitext erlauben',
+    freeTextExcerpt: 'Freitext',
+    freeTextCreatedAt: 'Eingegangen',
+    freeTextStatus: 'Öffentlich sichtbar',
+    freeTextOpenOverlay: 'Volltext öffnen',
+    freeTextOverlayText: 'Volltext',
+    freeTextOverlayStatus: 'Status',
+    freeTextOverlayCreatedAt: 'Zeitstempel',
+    summaryParticipationCount: 'Teilnahmen',
+    summarySubmissionCount: 'Abgaben',
+    summaryQuestionCount: 'Fragen',
+    summaryStatus: 'Status',
+    optionVotes: 'Stimmen',
+    optionPercentage: 'Anteil',
+    freeTextSection: 'Freitextantworten',
+    targetAreas: 'Zielgebiete',
+    targetAreasSearch: 'Zielgebiet suchen',
+    targetAreasSearchPlaceholder: 'Zielgebiet auswählen',
+    createdAt: 'Erstellt',
+    updatedAt: 'Aktualisiert',
+    publishedAt: 'Veröffentlicht',
+    archivedAt: 'Archiviert',
+    statusOptions: {
+      draft: 'Entwurf',
+      active: 'Aktiv',
+      archived: 'Archiviert',
+    },
+    resultVisibilityOptions: {
+      none: 'Nie',
+      afterSubmission: 'Nach Teilnahme',
+      afterSurveyEnd: 'Nach Umfrageende',
+    },
+    questionTypeOptions: {
+      singleChoice: 'Einfachauswahl',
+      multipleChoice: 'Mehrfachauswahl',
+      freeText: 'Freitext',
+      singleChoiceWithText: 'Einfachauswahl mit Freitext',
+      multipleChoiceWithText: 'Mehrfachauswahl mit Freitext',
+    },
+    freeTextStatusOptions: {
+      internal: 'Intern',
+      public: 'Öffentlich',
+    },
+  },
+} as const;
