+++ packages/plugin-surveys/src/plugin.translations.ts
@@ -1,534 +1,23 @@
+import { pluginSurveysContentEnTranslations } from './plugin.translations.content.en.js';
+import { pluginSurveysContentTranslations } from './plugin.translations.content.js';
+import { pluginSurveysMetaEnTranslations } from './plugin.translations.meta.en.js';
+import { pluginSurveysMetaTranslations } from './plugin.translations.meta.js';
+import { pluginSurveysStructureEnTranslations } from './plugin.translations.structure.en.js';
+import { pluginSurveysStructureTranslations } from './plugin.translations.structure.js';
+
 export const pluginSurveysTranslations = {
   de: {
     surveys: {
-      navigation: {
-        title: 'Umfragen',
-      },
-      actions: {
-        create: 'Umfrage anlegen',
-        edit: 'Umfrage ansehen',
-        update: 'Umfrage speichern',
-        delete: 'Umfrage löschen',
-        moderate: 'Freitexte moderieren',
-        export: 'Ergebnisse exportieren',
-        addTargetArea: 'Zielgebiet hinzufügen',
-        removeTargetArea: 'Zielgebiet {{name}} entfernen',
-        addQuestion: 'Frage hinzufügen',
-        addOption: 'Option hinzufügen',
-        deleteFreeText: 'Antwort {{index}} löschen',
-        moveQuestionUp: 'Frage {{index}} nach oben',
-        moveQuestionDown: 'Frage {{index}} nach unten',
-        deleteQuestion: 'Frage {{index}} löschen',
-        moveOptionUp: 'Option {{index}} nach oben',
-        moveOptionDown: 'Option {{index}} nach unten',
-        deleteOption: 'Option {{index}} löschen',
-        closeOverlay: 'Schließen',
-        confirmDelete: 'Löschen',
-        cancelDelete: 'Abbrechen',
-      },
-      pages: {
-        createTitle: 'Umfrage anlegen',
-        editTitle: 'Umfrage bearbeiten',
-        createDescription:
-          'Die Umfrage wird im normalen Content-Flow vorbereitet und nutzt bereits den stabilen Editor-Rahmen.',
-        editDescription:
-          'Die Umfrage ist in die normale Content-Struktur eingebunden und nutzt denselben Editor-Rahmen wie im Create-Fall.',
-      },
-      tabs: {
-        ariaLabel: 'Umfrage-Bereiche',
-        basis: {
-          label: 'Basis',
-          title: 'Basis',
-          description: 'Administrativer Rahmen der Umfrage.',
-        },
-        content: {
-          label: 'Inhalt',
-          title: 'Inhalt',
-          description: 'Redaktioneller Survey-Inhalt und Fragen.',
-        },
-        moderation: {
-          label: 'Moderation',
-          title: 'Moderation',
-          description: 'Freitext-Freigaben und Moderation.',
-        },
-        results: {
-          label: 'Ergebnisse',
-          title: 'Ergebnisse',
-          description: 'Überblick, Auswertung und Export.',
-        },
-        history: {
-          label: 'Historie',
-          title: 'Historie',
-          description: 'Änderungsverlauf der Umfrage.',
-        },
-      },
-      cards: {
-        basis: {
-          title: 'Basis',
-          description: 'Status, Laufzeit, Zielgebiet und Metadaten der Umfrage.',
-          identity: {
-            title: 'Identität',
-            description: 'Titel und Status der Umfrage.',
-          },
-          schedule: {
-            title: 'Laufzeit',
-            description: 'Start- und Endzeitraum der Umfrage.',
-          },
-          targetArea: {
-            title: 'Zielgebiet',
-            description: 'Optionale Zielgebiete für die Umfrage.',
-          },
-          metadata: {
-            title: 'Metadaten',
-            description: 'Zeitliche Metadaten der Umfrage.',
-          },
-        },
-        content: {
-          title: 'Inhalts-Rahmen',
-          description: 'Beschreibung, Hinweise und Frageneditor der Umfrage.',
-          descriptions: {
-            title: 'Beschreibung',
-            description: 'Kurz- und Langbeschreibung der Umfrage.',
-          },
-          participation: {
-            title: 'Teilnahme und Sichtbarkeit',
-            description: 'Teilnahmeoptionen und Sichtbarkeit der Ergebnisse.',
-          },
-          notices: {
-            title: 'Hinweise',
-            description: 'Datenschutz- und Transparenzhinweise.',
-          },
-          questions: {
-            title: 'Fragen',
-            description: 'Fragen und Antwortoptionen der Umfrage.',
-          },
-        },
-        moderation: {
-          title: 'Moderations-Rahmen',
-          description: 'Freitext-Freigaben werden nach dem ersten Speichern verfügbar.',
-        },
-        results: {
-          title: 'Ergebnis-Rahmen',
-          description: 'Ergebnisse und Exporte der Umfrage.',
-          summary: {
-            title: 'Übersicht',
-            description: 'Kompakter Überblick über die laufende Umfrage.',
-          },
-          questions: {
-            title: 'Frageergebnisse',
-            description: 'Aggregierte Ergebnisse pro Frage.',
-          },
-          export: {
-            title: 'Export',
-            description: 'Interne Exporte der Survey-Ergebnisse.',
-          },
-        },
-        history: {
-          title: 'Historien-Rahmen',
-          description: 'Historieneinträge werden nach dem ersten Speichern verfügbar.',
-        },
-      },
-      messages: {
-        createPendingHint:
-          'Dieser Bereich ist bereits sichtbar, wird aber erst nach dem ersten Speichern mit Daten gefüllt.',
-        sectionPlaceholder:
-          'Die fachlichen Felder dieses Bereichs folgen in den nächsten Umsetzungsabschnitten.',
-        historyPlaceholder: 'Die Historie erscheint hier, sobald die Umfrage bereits angelegt wurde.',
-        unlimitedScheduleHint: 'Ohne Enddatum bleibt die Umfrage unbefristet.',
-        targetAreasEmpty: 'Es stehen derzeit keine Zielgebiete zur Auswahl.',
-        metadataCreateHint: 'Metadaten erscheinen nach dem ersten Speichern der Umfrage.',
-        freeTextQuestionHint: 'Freitextfragen benötigen keine Antwortoptionen.',
-        optionSectionHint: 'Antwortoptionen bleiben als flache Abschnitte innerhalb der Frage.',
-        moderationEmpty: 'Für diese Umfrage liegen derzeit keine Freitextantworten zur Moderation vor.',
-        moderationEmptyQuestion: 'Zu dieser Frage liegen derzeit keine Freitextantworten vor.',
-        deleteQuestionTitle: 'Frage löschen',
-        deleteQuestionDescription: 'Soll diese Frage wirklich gelöscht werden?',
-        deleteOptionTitle: 'Option löschen',
-        deleteOptionDescription: 'Soll diese Antwortoption wirklich gelöscht werden?',
-        deleteFreeTextTitle: 'Freitextantwort löschen',
-        deleteFreeTextDescription: 'Soll diese Freitextantwort wirklich gelöscht werden?',
-        resultsQuestionEmpty: 'Für diese Frage liegen derzeit keine Ergebnisse vor.',
-        resultsFreeTextEmpty: 'Für diese Frage liegen keine Freitextantworten vor.',
-        resultsExportWithoutFreeText: 'Ergebnisexport ohne Freitextantworten.',
-        resultsExportWithFreeText: 'Ergebnisexport inklusive Freitextantworten.',
-      },
-      validation: {
-        titleRequired: 'Bitte einen Titel angeben.',
-        questionTitleRequired: 'Bitte einen Fragetitel angeben.',
-        optionTitleRequired: 'Bitte einen Optionstitel angeben.',
-      },
-      fields: {
-        title: 'Titel',
-        shortDescription: 'Kurzbeschreibung',
-        description: 'Beschreibung',
-        status: 'Status',
-        startAt: 'Start',
-        endAt: 'Ende',
-        isAnonymous: 'Anonyme Teilnahme',
-        showResultsInApp: 'Ergebnisse in der App anzeigen',
-        resultVisibility: 'Ergebnisfreigabe',
-        privacyNotice: 'Datenschutzhinweis',
-        transparencyNotice: 'Transparenzhinweis',
-        questionTitle: 'Fragetitel',
-        questionDescription: 'Fragebeschreibung',
-        questionType: 'Fragetyp',
-        questionRequired: 'Pflichtfrage',
-        optionTitle: 'Optionstitel',
-        optionEnablesFreeText: 'Freitext erlauben',
-        freeTextExcerpt: 'Freitext',
-        freeTextCreatedAt: 'Eingegangen',
-        freeTextStatus: 'Öffentlich sichtbar',
-        freeTextOpenOverlay: 'Volltext öffnen',
-        freeTextOverlayText: 'Volltext',
-        freeTextOverlayStatus: 'Status',
-        freeTextOverlayCreatedAt: 'Zeitstempel',
-        summaryParticipationCount: 'Teilnahmen',
-        summarySubmissionCount: 'Abgaben',
-        summaryQuestionCount: 'Fragen',
-        summaryStatus: 'Status',
-        optionVotes: 'Stimmen',
-        optionPercentage: 'Anteil',
-        freeTextSection: 'Freitextantworten',
-        targetAreas: 'Zielgebiete',
-        targetAreasSearch: 'Zielgebiet suchen',
-        targetAreasSearchPlaceholder: 'Zielgebiet auswählen',
-        createdAt: 'Erstellt',
-        updatedAt: 'Aktualisiert',
-        publishedAt: 'Veröffentlicht',
-        archivedAt: 'Archiviert',
-        statusOptions: {
-          draft: 'Entwurf',
-          active: 'Aktiv',
-          archived: 'Archiviert',
-        },
-        resultVisibilityOptions: {
-          none: 'Nie',
-          afterSubmission: 'Nach Teilnahme',
-          afterSurveyEnd: 'Nach Umfrageende',
-        },
-        questionTypeOptions: {
-          singleChoice: 'Einfachauswahl',
-          multipleChoice: 'Mehrfachauswahl',
-          freeText: 'Freitext',
-          singleChoiceWithText: 'Einfachauswahl mit Freitext',
-          multipleChoiceWithText: 'Mehrfachauswahl mit Freitext',
-        },
-        freeTextStatusOptions: {
-          internal: 'Intern',
-          public: 'Öffentlich',
-        },
-      },
-      labels: {
-        questionSection: 'Frage {{index}}',
-        optionSection: 'Option {{index}}',
-        freeTextVisibility: 'Antwort {{index}} öffentlich sichtbar',
-        exportWithoutFreeText: 'Export ohne Freitexte',
-        exportWithFreeText: 'Export mit Freitexten',
-        exportFormats: {
-          csv: 'CSV',
-          json: 'JSON',
-          excel: 'Excel',
-          xml: 'XML',
-        },
-      },
-      permissions: {
-        read: 'Umfragen lesen',
-        create: 'Umfragen anlegen',
-        update: 'Umfragen bearbeiten',
-        delete: 'Umfragen löschen',
-        moderate: 'Freitexte moderieren',
-        export: 'Ergebnisse exportieren',
-      },
-      history: {
-        createHint: 'Die Historie wird nach dem ersten Speichern verfügbar.',
-        loading: 'Historie wird geladen.',
-        empty: 'Noch keine Historie vorhanden.',
-        emptySummary: 'Keine zusätzliche Zusammenfassung vorhanden.',
-        tableLabel: 'Historie der Umfrage',
-        columns: {
-          time: 'Zeitpunkt',
-          action: 'Aktion',
-          actor: 'Akteur',
-          summary: 'Zusammenfassung',
-        },
-        errors: {
-          forbidden: 'Die Historie darf nicht angezeigt werden.',
-          notFound: 'Für diese Umfrage wurde keine Historie gefunden.',
-          load: 'Die Historie konnte nicht geladen werden.',
-        },
-        changedFields: 'Geänderte Felder: {{fields}}',
-        actions: {
-          created: 'Erstellt',
-          updated: 'Aktualisiert',
-          statusChanged: 'Status geändert',
-        },
-      },
+      ...pluginSurveysStructureTranslations,
+      ...pluginSurveysContentTranslations,
+      ...pluginSurveysMetaTranslations,
     },
   },
   en: {
     surveys: {
-      navigation: {
-        title: 'Surveys',
-      },
-      actions: {
-        create: 'Create survey',
-        edit: 'View survey',
-        update: 'Save survey',
-        delete: 'Delete survey',
-        moderate: 'Moderate free text',
-        export: 'Export results',
-        addTargetArea: 'Add target area',
-        removeTargetArea: 'Remove target area {{name}}',
-        addQuestion: 'Add question',
-        addOption: 'Add option',
-        deleteFreeText: 'Delete answer {{index}}',
-        moveQuestionUp: 'Move question {{index}} up',
-        moveQuestionDown: 'Move question {{index}} down',
-        deleteQuestion: 'Delete question {{index}}',
-        moveOptionUp: 'Move option {{index}} up',
-        moveOptionDown: 'Move option {{index}} down',
-        deleteOption: 'Delete option {{index}}',
-        closeOverlay: 'Close',
-        confirmDelete: 'Delete',
-        cancelDelete: 'Cancel',
-      },
-      pages: {
-        createTitle: 'Create survey',
-        editTitle: 'Edit survey',
-        createDescription:
-          'The survey is prepared in the standard content flow and already uses the stable editor frame.',
-        editDescription:
-          'The survey is already part of the standard content flow and uses the same editor frame as the create flow.',
-      },
-      tabs: {
-        ariaLabel: 'Survey sections',
-        basis: {
-          label: 'Basics',
-          title: 'Basics',
-          description: 'Administrative survey frame.',
-        },
-        content: {
-          label: 'Content',
-          title: 'Content',
-          description: 'Editorial survey content and questions.',
-        },
-        moderation: {
-          label: 'Moderation',
-          title: 'Moderation',
-          description: 'Free-text approvals and moderation.',
-        },
-        results: {
-          label: 'Results',
-          title: 'Results',
-          description: 'Overview, evaluation, and export.',
-        },
-        history: {
-          label: 'History',
-          title: 'History',
-          description: 'Survey change history.',
-        },
-      },
-      cards: {
-        basis: {
-          title: 'Basics',
-          description: 'Status, schedule, target area, and survey metadata.',
-          identity: {
-            title: 'Identity',
-            description: 'Survey title and status.',
-          },
-          schedule: {
-            title: 'Schedule',
-            description: 'Survey start and end window.',
-          },
-          targetArea: {
-            title: 'Target area',
-            description: 'Optional target areas for the survey.',
-          },
-          metadata: {
-            title: 'Metadata',
-            description: 'Temporal survey metadata.',
-          },
-        },
-        content: {
-          title: 'Content frame',
-          description: 'Description, notices, and the survey question editor.',
-          descriptions: {
-            title: 'Description',
-            description: 'Short and long survey description.',
-          },
-          participation: {
-            title: 'Participation and visibility',
-            description: 'Participation options and result visibility.',
-          },
-          notices: {
-            title: 'Notices',
-            description: 'Privacy and transparency notices.',
-          },
-          questions: {
-            title: 'Questions',
-            description: 'Survey questions and answer options.',
-          },
-        },
-        moderation: {
-          title: 'Moderation frame',
-          description: 'Free-text approvals become available after the first save.',
-        },
-        results: {
-          title: 'Results frame',
-          description: 'Survey results and exports.',
-          summary: {
-            title: 'Overview',
-            description: 'Compact overview of the running survey.',
-          },
-          questions: {
-            title: 'Question results',
-            description: 'Aggregated results per question.',
-          },
-          export: {
-            title: 'Export',
-            description: 'Internal exports of the survey results.',
-          },
-        },
-        history: {
-          title: 'History frame',
-          description: 'History entries become available after the first save.',
-        },
-      },
-      messages: {
-        createPendingHint:
-          'This section is already visible, but it will only be populated with data after the first save.',
-        sectionPlaceholder:
-          'The dedicated fields for this section follow in the next implementation steps.',
-        historyPlaceholder: 'The history will appear here once the survey has been created.',
-        unlimitedScheduleHint: 'Without an end date, the survey remains open-ended.',
-        targetAreasEmpty: 'There are currently no target areas available.',
-        metadataCreateHint: 'Metadata appears after the survey has been saved for the first time.',
-        freeTextQuestionHint: 'Free-text questions do not require answer options.',
-        optionSectionHint: 'Answer options remain flat sections inside the question.',
-        moderationEmpty: 'There are currently no free-text responses to moderate for this survey.',
-        moderationEmptyQuestion: 'There are currently no free-text responses for this question.',
-        deleteQuestionTitle: 'Delete question',
-        deleteQuestionDescription: 'Do you really want to delete this question?',
-        deleteOptionTitle: 'Delete option',
-        deleteOptionDescription: 'Do you really want to delete this answer option?',
-        deleteFreeTextTitle: 'Delete free-text response',
-        deleteFreeTextDescription: 'Do you really want to delete this free-text response?',
-        resultsQuestionEmpty: 'There are currently no results for this question.',
-        resultsFreeTextEmpty: 'There are no free-text responses for this question.',
-        resultsExportWithoutFreeText: 'Result export without free-text responses.',
-        resultsExportWithFreeText: 'Result export including free-text responses.',
-      },
-      validation: {
-        titleRequired: 'Please provide a title.',
-        questionTitleRequired: 'Please provide a question title.',
-        optionTitleRequired: 'Please provide an option title.',
-      },
-      fields: {
-        title: 'Title',
-        shortDescription: 'Short description',
-        description: 'Description',
-        status: 'Status',
-        startAt: 'Start',
-        endAt: 'End',
-        isAnonymous: 'Anonymous participation',
-        showResultsInApp: 'Show results in app',
-        resultVisibility: 'Result visibility',
-        privacyNotice: 'Privacy notice',
-        transparencyNotice: 'Transparency notice',
-        questionTitle: 'Question title',
-        questionDescription: 'Question description',
-        questionType: 'Question type',
-        questionRequired: 'Required question',
-        optionTitle: 'Option title',
-        optionEnablesFreeText: 'Allow free text',
-        freeTextExcerpt: 'Free text',
-        freeTextCreatedAt: 'Received',
-        freeTextStatus: 'Publicly visible',
-        freeTextOpenOverlay: 'Open full text',
-        freeTextOverlayText: 'Full text',
-        freeTextOverlayStatus: 'Status',
-        freeTextOverlayCreatedAt: 'Timestamp',
-        summaryParticipationCount: 'Participations',
-        summarySubmissionCount: 'Submissions',
-        summaryQuestionCount: 'Questions',
-        summaryStatus: 'Status',
-        optionVotes: 'Votes',
-        optionPercentage: 'Share',
-        freeTextSection: 'Free-text responses',
-        targetAreas: 'Target areas',
-        targetAreasSearch: 'Search target area',
-        targetAreasSearchPlaceholder: 'Select target area',
-        createdAt: 'Created',
-        updatedAt: 'Updated',
-        publishedAt: 'Published',
-        archivedAt: 'Archived',
-        statusOptions: {
-          draft: 'Draft',
-          active: 'Active',
-          archived: 'Archived',
-        },
-        resultVisibilityOptions: {
-          none: 'Never',
-          afterSubmission: 'After submission',
-          afterSurveyEnd: 'After survey end',
-        },
-        questionTypeOptions: {
-          singleChoice: 'Single choice',
-          multipleChoice: 'Multiple choice',
-          freeText: 'Free text',
-          singleChoiceWithText: 'Single choice with free text',
-          multipleChoiceWithText: 'Multiple choice with free text',
-        },
-        freeTextStatusOptions: {
-          internal: 'Internal',
-          public: 'Public',
-        },
-      },
-      labels: {
-        questionSection: 'Question {{index}}',
-        optionSection: 'Option {{index}}',
-        freeTextVisibility: 'Answer {{index}} publicly visible',
-        exportWithoutFreeText: 'Export without free texts',
-        exportWithFreeText: 'Export with free texts',
-        exportFormats: {
-          csv: 'CSV',
-          json: 'JSON',
-          excel: 'Excel',
-          xml: 'XML',
-        },
-      },
-      permissions: {
-        read: 'Read surveys',
-        create: 'Create surveys',
-        update: 'Update surveys',
-        delete: 'Delete surveys',
-        moderate: 'Moderate free text',
-        export: 'Export results',
-      },
-      history: {
-        createHint: 'The history becomes available after the first save.',
-        loading: 'History is loading.',
-        empty: 'No history available yet.',
-        emptySummary: 'No additional summary available.',
-        tableLabel: 'Survey history',
-        columns: {
-          time: 'Time',
-          action: 'Action',
-          actor: 'Actor',
-          summary: 'Summary',
-        },
-        errors: {
-          forbidden: 'The history cannot be displayed.',
-          notFound: 'No history was found for this survey.',
-          load: 'The history could not be loaded.',
-        },
-        changedFields: 'Changed fields: {{fields}}',
-        actions: {
-          created: 'Created',
-          updated: 'Updated',
-          statusChanged: 'Status changed',
-        },
-      },
+      ...pluginSurveysStructureEnTranslations,
+      ...pluginSurveysContentEnTranslations,
+      ...pluginSurveysMetaEnTranslations,
     },
   },
-} as const;
+};
