+++ packages/plugin-surveys/src/surveys.detail-card.tsx
@@ -1,0 +1,21 @@
+import type { ReactNode } from 'react';
+
+export function SurveyDetailCard({
+  title,
+  description,
+  children,
+}: Readonly<{
+  title: string;
+  description: string;
+  children: ReactNode;
+}>) {
+  return (
+    <section className="rounded-xl border border-border bg-card p-6 shadow-sm">
+      <div className="space-y-2">
+        <h3 className="text-base font-semibold text-foreground">{title}</h3>
+        <p className="text-sm leading-relaxed text-muted-foreground">{description}</p>
+      </div>
+      <div className="mt-5 border-t border-border pt-5">{children}</div>
+    </section>
+  );
+}
