+++ packages/plugin-surveys/src/surveys.detail-history-tab.tsx
@@ -3,6 +3,7 @@ import * as React from 'react';
 import { fetchIamContentHistory, formatDateTimeInEditorTimeZone } from '@sva/plugin-sdk';
 import { StudioLoadingState } from '@sva/studio-ui-react';
 
+import { SurveyDetailCard } from './surveys.detail-card.js';
 import { createSurveyHistoryEntries } from './surveys.history.js';
 
 export type SurveyDetailHistoryTabProps = Readonly<{
@@ -10,15 +11,11 @@ export type SurveyDetailHistoryTabProps = Readonly<{
   pt: (key: string, variables?: Readonly<Record<string, string | number>>) => string;
 }>;
 
-const resolveHistoryErrorMessage = (
-  pt: SurveyDetailHistoryTabProps['pt'],
-  error: unknown
-) => {
+const formatHistoryDate = (value: string) => formatDateTimeInEditorTimeZone(value) ?? value;
+
+const resolveHistoryErrorMessage = (pt: SurveyDetailHistoryTabProps['pt'], error: unknown) => {
   const code =
-    typeof error === 'object' &&
-    error !== null &&
-    'code' in error &&
-    typeof error.code === 'string'
+    typeof error === 'object' && error !== null && 'code' in error && typeof error.code === 'string'
       ? error.code
       : undefined;
 
@@ -32,8 +29,6 @@ const resolveHistoryErrorMessage = (
 
   return pt('history.errors.load');
 };
-
-const formatHistoryDate = (value: string) => formatDateTimeInEditorTimeZone(value) ?? value;
 
 const buildHistorySummary = (
   pt: SurveyDetailHistoryTabProps['pt'],
@@ -50,25 +45,7 @@ const buildHistorySummary = (
   return pt('history.emptySummary');
 };
 
-const SurveyHistoryCard = ({
-  title,
-  description,
-  children,
-}: Readonly<{
-  title: string;
-  description: string;
-  children: React.ReactNode;
-}>) => (
-  <section className="rounded-xl border border-border bg-card p-6 shadow-sm">
-    <div className="space-y-2">
-      <h3 className="text-base font-semibold text-foreground">{title}</h3>
-      <p className="text-sm leading-relaxed text-muted-foreground">{description}</p>
-    </div>
-    <div className="mt-5 border-t border-border pt-5">{children}</div>
-  </section>
-);
-
-export function SurveyDetailHistoryTab({ contentId, pt }: SurveyDetailHistoryTabProps) {
+function useSurveyHistoryState(contentId: string | undefined, pt: SurveyDetailHistoryTabProps['pt']) {
   const [entries, setEntries] = React.useState<ReturnType<typeof createSurveyHistoryEntries>>([]);
   const [isLoading, setIsLoading] = React.useState(Boolean(contentId));
   const [errorMessage, setErrorMessage] = React.useState<string | null>(null);
@@ -87,22 +64,19 @@ export function SurveyDetailHistoryTab({ contentId, pt }: SurveyDetailHistoryTab
 
     void fetchIamContentHistory(contentId)
       .then((historyEntries) => {
-        if (cancelled) {
-          return;
+        if (!cancelled) {
+          setEntries(createSurveyHistoryEntries(historyEntries));
         }
-        setEntries(createSurveyHistoryEntries(historyEntries));
       })
       .catch((error) => {
-        if (cancelled) {
-          return;
+        if (!cancelled) {
+          setErrorMessage(resolveHistoryErrorMessage(pt, error));
         }
-        setErrorMessage(resolveHistoryErrorMessage(pt, error));
       })
       .finally(() => {
-        if (cancelled) {
-          return;
+        if (!cancelled) {
+          setIsLoading(false);
         }
-        setIsLoading(false);
       });
 
     return () => {
@@ -110,63 +84,72 @@ export function SurveyDetailHistoryTab({ contentId, pt }: SurveyDetailHistoryTab
     };
   }, [contentId, pt]);
 
+  return { entries, errorMessage, isLoading };
+}
+
+function SurveyHistoryTable({ entries, pt }: Readonly<{ entries: ReturnType<typeof createSurveyHistoryEntries>; pt: SurveyDetailHistoryTabProps['pt'] }>) {
+  return (
+    <div className="overflow-x-auto">
+      <table className="min-w-full border-collapse" aria-label={pt('history.tableLabel')}>
+        <thead>
+          <tr className="border-b border-border/70 text-left text-sm">
+            <th className="px-3 py-2 font-semibold text-foreground">{pt('history.columns.time')}</th>
+            <th className="px-3 py-2 font-semibold text-foreground">{pt('history.columns.action')}</th>
+            <th className="px-3 py-2 font-semibold text-foreground">{pt('history.columns.actor')}</th>
+            <th className="px-3 py-2 font-semibold text-foreground">{pt('history.columns.summary')}</th>
+          </tr>
+        </thead>
+        <tbody>
+          {entries.map((entry) => (
+            <tr key={entry.id} className="border-b border-border/50 align-top text-sm last:border-b-0">
+              <td className="px-3 py-3 text-muted-foreground">{formatHistoryDate(entry.createdAt)}</td>
+              <td className="px-3 py-3 text-foreground">{pt(entry.actionLabelKey)}</td>
+              <td className="px-3 py-3 text-muted-foreground">{entry.actor}</td>
+              <td className="px-3 py-3 text-muted-foreground">{buildHistorySummary(pt, entry)}</td>
+            </tr>
+          ))}
+        </tbody>
+      </table>
+    </div>
+  );
+}
+
+export function SurveyDetailHistoryTab({ contentId, pt }: SurveyDetailHistoryTabProps) {
+  const { entries, errorMessage, isLoading } = useSurveyHistoryState(contentId, pt);
+
   if (!contentId) {
     return (
-      <SurveyHistoryCard title={pt('cards.history.title')} description={pt('cards.history.description')}>
+      <SurveyDetailCard title={pt('cards.history.title')} description={pt('cards.history.description')}>
         <p className="text-sm text-muted-foreground">{pt('history.createHint')}</p>
-      </SurveyHistoryCard>
+      </SurveyDetailCard>
     );
   }
 
   if (isLoading) {
     return (
-      <SurveyHistoryCard title={pt('cards.history.title')} description={pt('cards.history.description')}>
+      <SurveyDetailCard title={pt('cards.history.title')} description={pt('cards.history.description')}>
         <StudioLoadingState>{pt('history.loading')}</StudioLoadingState>
-      </SurveyHistoryCard>
+      </SurveyDetailCard>
     );
   }
 
   if (errorMessage) {
     return (
-      <SurveyHistoryCard title={pt('cards.history.title')} description={pt('cards.history.description')}>
-        <div
-          role="alert"
-          className="rounded-lg border border-destructive/30 bg-destructive/5 p-4 text-sm text-destructive"
-        >
+      <SurveyDetailCard title={pt('cards.history.title')} description={pt('cards.history.description')}>
+        <div role="alert" className="rounded-lg border border-destructive/30 bg-destructive/5 p-4 text-sm text-destructive">
           {errorMessage}
         </div>
-      </SurveyHistoryCard>
+      </SurveyDetailCard>
     );
   }
 
   return (
-    <SurveyHistoryCard title={pt('cards.history.title')} description={pt('cards.history.description')}>
+    <SurveyDetailCard title={pt('cards.history.title')} description={pt('cards.history.description')}>
       {entries.length === 0 ? (
         <p className="text-sm text-muted-foreground">{pt('history.empty')}</p>
       ) : (
-        <div className="overflow-x-auto">
-          <table className="min-w-full border-collapse" aria-label={pt('history.tableLabel')}>
-            <thead>
-              <tr className="border-b border-border/70 text-left text-sm">
-                <th className="px-3 py-2 font-semibold text-foreground">{pt('history.columns.time')}</th>
-                <th className="px-3 py-2 font-semibold text-foreground">{pt('history.columns.action')}</th>
-                <th className="px-3 py-2 font-semibold text-foreground">{pt('history.columns.actor')}</th>
-                <th className="px-3 py-2 font-semibold text-foreground">{pt('history.columns.summary')}</th>
-              </tr>
-            </thead>
-            <tbody>
-              {entries.map((entry) => (
-                <tr key={entry.id} className="border-b border-border/50 align-top text-sm last:border-b-0">
-                  <td className="px-3 py-3 text-muted-foreground">{formatHistoryDate(entry.createdAt)}</td>
-                  <td className="px-3 py-3 text-foreground">{pt(entry.actionLabelKey)}</td>
-                  <td className="px-3 py-3 text-muted-foreground">{entry.actor}</td>
-                  <td className="px-3 py-3 text-muted-foreground">{buildHistorySummary(pt, entry)}</td>
-                </tr>
-              ))}
-            </tbody>
-          </table>
-        </div>
+        <SurveyHistoryTable entries={entries} pt={pt} />
       )}
-    </SurveyHistoryCard>
+    </SurveyDetailCard>
   );
 }
